organize validator class
