+++ src/main/java/com/clivern/kangaroo/constraint/ConstraintType.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright (C) 2019 Clivern <http://clivern.com>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package com.clivern.kangaroo.constraint;
+
+/** ConstraintType Class */
+public enum ConstraintType {
+    OBJECT("OBJECT"),
+    INTEGER("INTEGER"),
+    STRING("STRING"),
+    ARRAY("ARRAY");
+
+    private final String name;
+
+    /**
+     * Class Constructor
+     *
+     * @param s the string value
+     */
+    private ConstraintType(String s) {
+        name = s;
+    }
+
+    /**
+     * Compare to string
+     *
+     * @param otherName the string value to compare against
+     * @return whether value equals the argument
+     */
+    public boolean equalsName(String otherName) {
+        return name.equals(otherName);
+    }
+
+    /**
+     * Change to string
+     *
+     * @return the string value
+     */
+    public String toString() {
+        return this.name;
+    }
+}
